audit/requirements.txt
# Core dependencies
qdrant-client>=1.6.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# OpenAI and LLM
openai>=1,<2

# Environment and configuration
python-dotenv>=1,<2

# Data processing
orjson>=3,<4
tqdm>=4,<5
rapidfuzz>=3,<4

# Web interface
streamlit>=1.28.0

# Embedding models
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.21.0

# Standard library (no need to specify)
# uuid

